fix vesting

+++ contracts/validator/DTOStaking.sol
@@ -15,7 +15,13 @@ import "@openzeppelin/contracts-upgradeable/security/PausableUpgradeable.sol";
 import "../interfaces/IStakingTokenLock.sol";
 
 // https://docs.synthetix.io/contracts/source/contracts/stakingrewards
-contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeable,PausableUpgradeable,OwnableUpgradeable{
+contract DTOStaking is
+    Initializable,
+    ReentrancyGuardUpgradeable,
+    UUPSUpgradeable,
+    PausableUpgradeable,
+    OwnableUpgradeable
+{
     using SafeMathUpgradeable for uint256;
     using SafeERC20Upgradeable for IERC20Upgradeable;
 
@@ -37,67 +43,84 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
     uint256 public stakingLockTime;
 
     address public rewardsDistribution;
-     function initialize(
+
+    function initialize(
         address _rewardsToken,
         address _stakingToken,
         address _stakingTokenLock,
-        uint256  _periodFinish ,
-        uint256  _rewardRate ,
-        uint256  _rewardsDuration,
+        uint256 _periodFinish,
+        uint256 _rewardRate,
+        uint256 _rewardsDuration,
         uint256 _stakingLockTime
-
     ) public initializer {
         __Ownable_init();
         rewardsToken = IERC20Upgradeable(_rewardsToken);
         stakingToken = IERC20Upgradeable(_stakingToken);
         stakingTokenLock = IStakingTokenLock(_stakingTokenLock);
         periodFinish = _periodFinish;
-        rewardRate  =_rewardRate;
-        rewardsDuration =  _rewardsDuration;
+        rewardRate = _rewardRate;
+        rewardsDuration = _rewardsDuration;
         stakingLockTime = _stakingLockTime;
     }
 
     /* ========== CONSTRUCTOR ========== */
-        /// @custom:oz-upgrades-unsafe-allow constructor
-   constructor() initializer {}
-   function _authorizeUpgrade(address) internal override onlyOwner {}
+    /// @custom:oz-upgrades-unsafe-allow constructor
+    constructor() initializer {}
+
+    function _authorizeUpgrade(address) internal override onlyOwner {}
+
     /* ========== VIEWS ========== */
 
-    function totalSupply() external  view returns (uint256) {
+    function totalSupply() external view returns (uint256) {
         return _totalSupply;
     }
 
-    function balanceOf(address account) external  view returns (uint256) {
+    function balanceOf(address account) external view returns (uint256) {
         return _balances[account];
     }
 
-    function lastTimeRewardApplicable() public  view returns (uint256) {
+    function lastTimeRewardApplicable() public view returns (uint256) {
         return block.timestamp < periodFinish ? block.timestamp : periodFinish;
     }
 
-    function rewardPerToken() public  view returns (uint256) {
+    function rewardPerToken() public view returns (uint256) {
         if (_totalSupply == 0) {
             return rewardPerTokenStored;
         }
         return
             rewardPerTokenStored.add(
-                lastTimeRewardApplicable().sub(lastUpdateTime).mul(rewardRate).mul(1e18).div(_totalSupply)
+                lastTimeRewardApplicable()
+                    .sub(lastUpdateTime)
+                    .mul(rewardRate)
+                    .mul(1e18)
+                    .div(_totalSupply)
             );
     }
 
-    function earned(address account) public  view returns (uint256) {
-        return _balances[account].mul(rewardPerToken().sub(userRewardPerTokenPaid[account])).div(1e18).add(rewards[account]);
+    function earned(address account) public view returns (uint256) {
+        return
+            _balances[account]
+                .mul(rewardPerToken().sub(userRewardPerTokenPaid[account]))
+                .div(1e18)
+                .add(rewards[account]);
     }
 
-    function getRewardForDuration() external  view returns (uint256) {
+    function getRewardForDuration() external view returns (uint256) {
         return rewardRate.mul(rewardsDuration);
     }
+
     function changeLockTime(uint256 _stakingLockTime) external onlyOwner {
         stakingLockTime = _stakingLockTime;
     }
+
     /* ========== MUTATIVE FUNCTIONS ========== */
 
-    function stake(uint256 amount) external nonReentrant whenNotPaused updateReward(msg.sender) {
+    function stake(uint256 amount)
+        external
+        nonReentrant
+        whenNotPaused
+        updateReward(msg.sender)
+    {
         require(amount > 0, "Cannot stake 0");
         _totalSupply = _totalSupply.add(amount);
         _balances[msg.sender] = _balances[msg.sender].add(amount);
@@ -105,24 +128,32 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
         emit Staked(msg.sender, amount);
     }
 
-    function withdraw(uint256 amount) public nonReentrant updateReward(msg.sender) {
+    function withdraw(uint256 amount)
+        public
+        nonReentrant
+        updateReward(msg.sender)
+    {
         require(amount > 0, "Cannot withdraw 0");
         _totalSupply = _totalSupply.sub(amount);
         _balances[msg.sender] = _balances[msg.sender].sub(amount);
         stakingToken.safeApprove(address(stakingTokenLock), amount);
-        stakingTokenLock.lock(address(stakingToken), msg.sender, amount, stakingLockTime);
+        stakingTokenLock.lock(
+            address(stakingToken),
+            msg.sender,
+            amount,
+            stakingLockTime
+        );
 
         emit Withdrawn(msg.sender, amount);
     }
 
-    
     function getReward() public nonReentrant updateReward(msg.sender) {
         uint256 reward = rewards[msg.sender];
         if (reward > 0) {
             rewards[msg.sender] = 0;
             rewardsToken.safeTransfer(msg.sender, reward);
             emit RewardPaid(msg.sender, reward);
-        } 
+        }
     }
 
     function exit() external {
@@ -131,8 +162,12 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
     }
 
     /* ========== RESTRICTED FUNCTIONS ========== */
-   
-    function notifyRewardAmount(uint256 reward) external onlyOwner updateReward(address(0)) {
+
+    function notifyRewardAmount(uint256 reward)
+        external
+        onlyOwner
+        updateReward(address(0))
+    {
         if (block.timestamp >= periodFinish) {
             rewardRate = reward.div(rewardsDuration);
         } else {
@@ -145,8 +180,11 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
         // This keeps the reward rate in the right range, preventing overflows due to
         // very high values of rewardRate in the earned and rewardsPerToken functions;
         // Reward + leftover must be less than 2^256 / 10^18 to avoid overflow.
-        uint balance = rewardsToken.balanceOf(address(this));
-        require(rewardRate <= balance.div(rewardsDuration), "Provided reward too high");
+        uint256 balance = rewardsToken.balanceOf(address(this));
+        require(
+            rewardRate <= balance.div(rewardsDuration),
+            "Provided reward too high"
+        );
 
         lastUpdateTime = block.timestamp;
         periodFinish = block.timestamp.add(rewardsDuration);
@@ -154,8 +192,14 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
     }
 
     // Added to support recovering LP Rewards from other systems such as BAL to be distributed to holders
-    function recoverERC20(address tokenAddress, uint256 tokenAmount) external onlyOwner {
-        require(tokenAddress != address(stakingToken), "Cannot withdraw the staking token");
+    function recoverERC20(address tokenAddress, uint256 tokenAmount)
+        external
+        onlyOwner
+    {
+        require(
+            tokenAddress != address(stakingToken),
+            "Cannot withdraw the staking token"
+        );
         IERC20Upgradeable(tokenAddress).safeTransfer(msg.sender, tokenAmount);
         emit Recovered(tokenAddress, tokenAmount);
     }
@@ -180,21 +224,26 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
         }
         _;
     }
- /* ==========  ========== */
+
+    /* ==========  ========== */
 
     function getUnlock(address _addr, uint256 index) public {
-         stakingTokenLock.unlock(_addr,index);
+        stakingTokenLock.unlock(_addr, index);
     }
-    function getLockInfo(address _user)   external
+
+    function getLockInfo(address _user)
+        external
         view
         returns (
             bool[] memory isWithdrawns,
             address[] memory tokens,
             uint256[] memory unlockableAts,
             uint256[] memory amounts
-        ){
-            return stakingTokenLock.getLockInfo(_user);
-        }
+        )
+    {
+        return stakingTokenLock.getLockInfo(_user);
+    }
+
     function getLockInfoByIndexes(address _addr, uint256[] memory _indexes)
         external
         view
@@ -207,13 +256,10 @@ contract DTOStaking is Initializable, ReentrancyGuardUpgradeable, UUPSUpgradeabl
     {
         return stakingTokenLock.getLockInfoByIndexes(_addr, _indexes);
     }
-     function getLockInfoLength(address _addr)
-        external
-        view
-        returns (uint256)
-    {
+
+    function getLockInfoLength(address _addr) external view returns (uint256) {
         return stakingTokenLock.getLockInfoLength(_addr);
-    }    
+    }
 
     /* ========== EVENTS ========== */
